库模块已完善；故障管理已完善；代码生成已完善

+++ utilities/PrjScript/UnitDesign/NewUnitMdl.m
@@ -9,17 +9,24 @@ if isempty(unitName)
     return;
 end
 
+if exist([unitName '.slx'],'file')
+    disp('A model with the same name already Exists!');
+    return;
+end
+
 sfolder=fullfile(projectRoot, 'cm\Unitfolder');
 dfolder=pwd;
 
 copyfile(fullfile(sfolder,'Unit.slx'),fullfile(dfolder,[unitName '.slx']),'f');
-addFile(project,fullfile(dfolder,[unitName '.slx']));
+addFile(currentProject,fullfile(dfolder,[unitName '.slx']));
 
  % 关联模型和dd文件
  % Link the dictionary to the model 
  
  load_system([unitName '.slx']);
  set_param(unitName,'DataDictionary', [unitName '.sldd']);
+ 
+%  set_param(unitName,'EnableAccessToBaseWorkspace','on');
  
  % 配置模型ConfigSet
  cref = Simulink.ConfigSetRef;
